+++ packages/dsw-document-worker/dsw/document_worker/templates/steps/template.py
@@ -32,6 +32,10 @@ class Jinja2Step(Step):
     OPTION_ROOT_FILE = 'template'
     OPTION_CONTENT_TYPE = 'content-type'
     OPTION_EXTENSION = 'extension'
+    OPTION_JINJA_EXT = 'jinja-ext'
+    OPTION_I18N_DIR = 'i18n-dir'
+    OPTION_I18N_DOMAIN = 'i18n-domain'
+    OPTION_I18N_LANG = 'i18n-lang'
 
     def _jinja_exception_msg(self, e: jinja2.exceptions.TemplateSyntaxError):
         lines = [
@@ -47,19 +51,46 @@ class Jinja2Step(Step):
         self.root_file = self.options[self.OPTION_ROOT_FILE]
         self.content_type = self.options.get(self.OPTION_CONTENT_TYPE, self.DEFAULT_FORMAT.content_type)
         self.extension = self.options.get(self.OPTION_EXTENSION, self.DEFAULT_FORMAT.file_extension)
+        self.jinja_ext = frozenset(
+            map(lambda x: x.strip(), self.options.get(self.OPTION_JINJA_EXT, '').split(','))
+        )
+        self.i18n_dir = self.options.get(self.OPTION_I18N_DIR, None)
+        self.i18n_domain = self.options.get(self.OPTION_I18N_DOMAIN, 'default')
+        self.i18n_lang = self.options.get(self.OPTION_I18N_LANG, None)
 
         self.output_format = FileFormat(self.extension, self.content_type, self.extension)
         try:
             self.j2_env = jinja2.Environment(
                 loader=jinja2.FileSystemLoader(searchpath=template.template_dir),
-                extensions=['jinja2.ext.do'],
+                extensions=[
+                    'jinja2.ext.do',
+                    'jinja2.ext.loopcontrols',
+                ],
             )
+            if 'i18n' in self.jinja_ext:
+                self._add_j2_i18n(template)
+            if 'debug' in self.jinja_ext:
+                self.j2_env.add_extension('jinja2.ext.debug')
             self._add_j2_enhancements()
             self.j2_root_template = self.j2_env.get_template(self.root_file)
         except jinja2.exceptions.TemplateSyntaxError as e:
             self.raise_exc(self._jinja_exception_msg(e))
         except Exception as e:
             self.raise_exc(f'Failed loading Jinja2 template: {e}')
+
+    def _add_j2_i18n(self, template):
+        self.j2_env.add_extension('jinja2.ext.i18n')
+        if self.i18n_dir is not None and self.i18n_lang is not None:
+            import gettext
+            locale_path = template.template_dir / self.i18n_dir
+            translations = gettext.translation(
+                domain=self.i18n_domain,
+                localedir=locale_path,
+                languages=map(lambda x: x.strip(), self.i18n_lang.split(',')),
+            )
+            self.j2_env.install_gettext_translations(translations, newstyle=True)  # type: ignore
+        else:
+            self.j2_env.install_null_translations(newstyle=True)  # type: ignore
 
     def _add_j2_enhancements(self):
         from ..filters import filters