+++ packages/dsw-mailer/dsw/mailer/consts.py
@@ -1,5 +1,5 @@
 PROG_NAME = 'dsw-mailer'
-VERSION = '3.20.2'
+VERSION = '3.21.0'
 
 LOGGER_NAME = 'mailer'