+++ packages/dsw-mailer/CHANGELOG.md
@@ -8,6 +8,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 
+## [3.21.0]
+
+### Added
+
+- Add 2FA verification email
+- Support for email headers: `date`, `message-id`, `language`, `priority`, `sensitivity`, and `importance`
+
 ## [3.20.2]
 
 Released for version consistency with other DSW tools.
@@ -100,3 +107,4 @@ Released for version consistency with other DSW tools.
 [3.20.0]: /../../tree/v3.20.0
 [3.20.1]: /../../tree/v3.20.1
 [3.20.2]: /../../tree/v3.20.2
+[3.21.0]: /../../tree/v3.21.0