+++ packages/dsw-tdk/CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [3.21.0]
+
+### Changed
+
+- Rectify API URL (trailing slash)
+
 ## [3.20.2]
 
 ### Fixed
@@ -289,3 +295,4 @@ Initial DSW Template Development Kit (versioned as part of the [DSW platform](ht
 [3.20.0]: /../../tree/v3.20.0
 [3.20.1]: /../../tree/v3.20.1
 [3.20.2]: /../../tree/v3.20.2
+[3.21.0]: /../../tree/v3.21.0